Fix catch_error decorator
With the new changes in deepaas (predict will use
the same pool as train) the old version of the decorator
(which returned a aiohttp HTTPBadRequest exception) no
longer works (due to unpickling error, see [1]).

This new decorator will return a 200 response but with
a status:'error'. Not optimal but does the job for the
time being.

[1] https://bugs.python.org/issue39751